requirements.txt
selenium==4.38.0
beautifulsoup4==4.12.3
lxml==5.2.1
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
colorama==0.4.6
urllib3>=2.5.0,<3.0
requests==2.31.0